refactor(forum api): code style improvements (GAM-42)
