+++ packages/server/index.ts
@@ -5,26 +5,19 @@ import { Sequelize, SequelizeOptions } from 'sequelize-typescript';
 import { createServer as createViteServer, type ViteDevServer } from 'vite';
 
 import { router as apiRouter } from './routes/api';
+import { router as practicumRouter } from './routes/practicum';
 
 dotenv.config();
 
 import { configureStore } from '@reduxjs/toolkit';
 import express from 'express';
 import * as fs from 'fs';
-import https from 'https';
 import * as path from 'path';
 
 import { commentModel } from './models/comment';
 import { topicModel } from './models/topic';
 import { userModel } from './models/user';
 import { baseApi } from './store';
-
-type requestOptions = {
-    host: string;
-    path: string;
-    method?: string;
-    headers?: Record<string, string | number>;
-};
 
 const isDev = () => process.env.NODE_ENV === 'development';
 
@@ -75,42 +68,7 @@ async function startServer() {
     }
 
     app.use('/api', apiRouter);
-
-    app.use('/practicum/*', async (_req, _res) => {
-        const path = _req.baseUrl.replace('/practicum', '');
-
-        const options: requestOptions = {
-            host: 'ya-praktikum.tech',
-            path: '/api/v2' + path,
-            method: _req.method,
-            headers: {},
-        };
-        if (_req.headers?.cookie && options.headers) {
-            options.headers.cookie = _req.headers?.cookie;
-        }
-        if (_req.body && options.headers) {
-            options.headers['Content-Type'] = 'application/json';
-            options.headers['Content-Length'] = JSON.stringify(_req.body).length;
-        }
-        const req = https.request(options, (res) => {
-            const proxyCookies: string[] = [];
-            res.headers['set-cookie']?.forEach((e) => proxyCookies.push(e.replace('Domain=ya-praktikum.tech', '')));
-            _res.writeHead(200, {
-                'Set-Cookie': proxyCookies,
-            });
-            res.on('data', function (chunk) {
-                _res.status(res.statusCode || 400).end(chunk);
-            });
-        });
-
-        req.on('error', (e) => {
-            _res.status(400).end(JSON.stringify({ reason: e }));
-            console.error('Error: ' + e);
-        });
-
-        if (_req.body) req.write(JSON.stringify(_req.body));
-        req.end();
-    });
+    app.use('/practicum/*', practicumRouter);
 
     if (!isDev()) {
         app.use('/assets', express.static(path.resolve(distPath, 'assets')));
